No need to cast pointer to an integer.

+++ port/kernel/avr/avr.h
@@ -28,7 +28,7 @@
 #define AVR_HARD_RESET          FALSE
 #define HW_TICKS_PER_SEC        (PCLK_FREQ / 64)
 #define SYSTEM_STACK            (((uint8_t *)&__heap_start) + TARGET_HEAP_SIZE)
-#define SYS_STACK_SIZE          (system_stack_end - (uint32_t)SYSTEM_STACK)
+#define SYS_STACK_SIZE          (system_stack_end - SYSTEM_STACK)
 
 /* System interrupt level. */
 extern volatile uint8_t sys_interrupt_level;
